LatSpec.jl Documentation

LatSpec.autocor — Function

autocor(x,lag)

Returns autocorrelation function for a distance of lag, i.e. between values of x[t] and x[t+lag]. The autocorrelation function is normalized such that autocor(x,0) = 1. This is equvivalent to the quantity $\Gamma_X(t)$ of equation (4.61) in Gattringer/Lang (2010).

autocor(x)

Returns autocorrelation function for a set of lags from $0$ up to the closest integer to $10 \text{log}_{10}(l)$. Here $l$ is the number of elements in x.

LatSpec.autotimeexp — Function

autotimeexp(x)

Returns $\max(1,\tau)$ where $\tau$ is the exponential autocorrelation time of a series of measurements $x$. This is obtained by fitting the autocorrelation function to an exponential function of the form $A \exp(\frac{t}{\tau})$.