LatSpec.jl Documentation

LatSpec.DataPoint — Type

DataPoint(x,err)

Parametric type that stores a value and its respective statistical uncertainty. The uncertainty can either be a number for a symmetric uncertainty or a tuple for lower and upper bounds. It can be created by either providing positional arguments (e.g. DataPoint(1.0,0.1) or DataPoint(1.0,(0.09,0.11))) or by using the keyword arguments lower and upper (e.g. `DataPoint(1.0,lower=0.1,upper=0.2). If only one keyword is given, the other is assumed to be zero.

It can also be created by using the ± symbol, e.g. x ± err.

LatSpec.autocor — Method

autocor(x,lag)

Returns autocorrelation function for a distance of lag, i.e. between values of x[t] and x[t+lag]. The autocorrelation function is normalized such that autocor(x,0) = 1. This is equvivalent to the quantity $\Gamma_X(t)$ of equation (4.61) in Gattringer/Lang (2010).

LatSpec.autocor — Method

autocor(x)

Returns autocorrelation function for a set of lags from $0$ up to the closest integer to $10 \text{log}_{10}(l)$. Here $l$ is the number of elements in x.

LatSpec.autotimeexp — Method

autotimeexp(x)

Returns $\max(1,\tau)$ where $\tau$ is the exponential autocorrelation time of a series of measurements $x$. This is obtained by fitting the autocorrelation function to an exponential function of the form $A \exp(\frac{t}{\tau})$.

LatSpec.err — Method

err(x::DataPoint)

Returns the statistical uncertainty of the data point as a tuple of upper and lower bounds. If none is provided it will default to zero.

LatSpec.theories — Method

Returns a list of supported Theorys

LatSpec.theory — Method

Set the Spectroscopy Theory.

LatSpec.theory — Method

Returns the current theory used for spectroscopy. Initializes package on first call.

LatSpec.theory_name — Method

Returns the name of the current Theory

LatSpec.value — Method

value(x::DataPoint)

Returns the mean value of the data point.